// frontend/src/api/model/recentCompletionItem.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Whendoist
 * WHEN do I do my tasks?
 * OpenAPI spec version: 0.48.0
 */

export interface RecentCompletionItem {
  id: number;
  task_id: number;
  title: string;
  completed_at_display: string;
  domain_name: string;
  domain_icon: string;
  is_instance: boolean;
}